fix bug with firebase in client - move fetching functions to util to fix nextjs bug - changes details component to send undefined instead of "" with no file input - fixes a bug with text form data not being sent
