+++ src/lib/qbo-api.ts
@@ -1,7 +1,7 @@
 import { Session } from "next-auth"
 import { ParsedUrlQuery } from "querystring"
 
-import { fetchJsonData } from "./app-api"
+import { fetchJsonData } from "./util"
 import { DbUser } from "./db"
 import { formatDateHtmlReverse } from "./util"
