+++ src/lib/util.ts
@@ -36,3 +36,50 @@ export type DeepPartial<T> = T extends object
       [P in keyof T]?: DeepPartial<T[P]>
     }
   : T
+
+export const getUrl = () => {
+  const vercelUrl = process.env.NEXT_PUBLIC_VERCEL_URL
+  const url = vercelUrl ? `https://${vercelUrl}` : "http://localhost:3000/"
+  if (url.at(-1) === "/") return url
+  else return `${url}/`
+}
+
+export async function fetchJsonData<T = any>(url: string, accessToken?: string): Promise<T> {
+  const response = await (accessToken
+    ? fetch(url, {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+          Accept: "application/json",
+        },
+      })
+    : fetch(url, {
+        headers: {
+          Accept: "application/json",
+        },
+      }))
+
+  if (!response.ok) {
+    throw new Error(`GET request to url: ${url} failed`)
+  }
+
+  const report: T = await response.json()
+  return report
+}
+
+export async function postJsonData<T = any>(url: string, json: any): Promise<T> {
+  const response = await fetch(url, {
+    method: "POST",
+    headers: {
+      Accept: "application/json",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(json),
+  })
+
+  if (!response.ok) {
+    throw new Error(`POST request to url: ${url} failed`)
+  }
+
+  const report: T = await response.json()
+  return report
+}
