+++ src/pages/api/auth/[...nextauth].ts
@@ -4,7 +4,8 @@ import { OAuthConfig } from "next-auth/providers"
 
 import { user } from "@/lib/db"
 import { QBOProfile } from "@/lib/qbo-api"
-import { base64EncodeString, fetchJsonData } from "@/lib/app-api"
+import { base64EncodeString } from "@/lib/app-api"
+import { fetchJsonData } from "@/lib/util"
 
 const customProvider: OAuthConfig<QBOProfile> = {
   id: "QBO",
